feat: Implement Home Search Overlay and UI State Management
- Added HomeSearchResultsOverlay composable for displaying search results from Jellyfin and Seerr.
- Introduced HomeUiEvent sealed interface to handle various UI events related to home screen interactions.
- Created HomeUiState data class to manage the state of the home screen, including search state and results.
- Refactored HomeViewModel to utilize HomeUiState for better state management and event handling.
- Enhanced KidsHomeScreen to accept HomeUiState for improved data flow and UI updates.
