+++ feature/home/src/main/java/com/raulshma/jellyplay/feature/home/HomeViewModel.kt
@@ -1,9 +1,5 @@
 package com.raulshma.jellyplay.feature.home
 
-import androidx.compose.runtime.Immutable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ProcessLifecycleOwner
@@ -20,37 +16,32 @@ import com.raulshma.jellyplay.core.datastore.UserPreferencesStore
 import com.raulshma.jellyplay.core.model.seerr.DiscoverSectionType
 import com.raulshma.jellyplay.core.model.seerr.SeerrSearchResponse
 import com.raulshma.jellyplay.core.model.HomeMode
-import com.raulshma.jellyplay.core.model.HomeSection
 import com.raulshma.jellyplay.core.model.HomeSectionType
-import com.raulshma.jellyplay.core.model.MediaItem
-import com.raulshma.jellyplay.core.model.OfflineMediaItem
 import com.raulshma.jellyplay.core.model.OfflineMode
 import com.raulshma.jellyplay.core.model.seerr.SeerrPreferences
 import com.raulshma.jellyplay.core.model.seerr.SeerrSearchItem
-import com.raulshma.jellyplay.core.model.seerr.SeerrRadarrServiceDetail
-import com.raulshma.jellyplay.core.model.seerr.SeerrSonarrServiceDetail
-import com.raulshma.jellyplay.core.model.seerr.SeerrSeason
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.debounce
 import kotlinx.coroutines.flow.distinctUntilChanged
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.async
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.withContext
 import java.time.LocalDate
 import java.time.ZoneOffset
 import javax.inject.Inject
+import dagger.hilt.android.lifecycle.HiltViewModel
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
@@ -71,44 +62,16 @@ class HomeViewModel @Inject constructor(
         private const val MIN_REFRESH_INTERVAL_MS = 30_000L
     }
 
-    var sections by mutableStateOf<List<HomeSection>>(emptyList())
-        private set
-    var favorites by mutableStateOf<List<MediaItem>>(emptyList())
-        private set
-    var isLoading by mutableStateOf(true)
-        private set
-    var isRefreshing by mutableStateOf(false)
-        private set
-    var error by mutableStateOf<String?>(null)
-        private set
-    var kidsModeEnabled by mutableStateOf(false)
-        private set
-    var homeMode by mutableStateOf(HomeMode.VIDEO)
-        private set
-    var dynamicTheming by mutableStateOf(true)
-        private set
-    var oledMode by mutableStateOf(false)
-        private set
-    var offlineMode by mutableStateOf(OfflineMode.ONLINE)
-        private set
-    var offlineLibrary by mutableStateOf<List<OfflineMediaItem>>(emptyList())
-        private set
-    var enabledHomeSectionTypes by mutableStateOf(HomeSectionType.CONFIGURABLE)
-        private set
-    var hiddenLibrarySectionIds by mutableStateOf(emptySet<String>())
-        private set
-
-    // Seerr Discover state
-    var discoverSections by mutableStateOf<Map<DiscoverSectionType, List<SeerrSearchItem>>>(emptyMap())
-        private set
-    var discoverEnabled by mutableStateOf(false)
-        private set
-    private var seerrPreferences by mutableStateOf(SeerrPreferences())
-
-    private var lastContinueWatchingIds: Set<String> = emptySet()
+    private val _uiState = MutableStateFlow(HomeUiState())
+    val uiState: StateFlow<HomeUiState> = _uiState.asStateFlow()
 
     val activeDownloadCount = downloadRepository.getActiveDownloadCount()
         .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5_000), 0)
+
+    private var enabledHomeSectionTypes = HomeSectionType.CONFIGURABLE
+    private var hiddenLibrarySectionIds = emptySet<String>()
+    private var lastContinueWatchingIds: Set<String> = emptySet()
+    private var seerrPreferences = SeerrPreferences()
 
     private val refreshMutex = Mutex()
     private var refreshJob: Job? = null
@@ -117,50 +80,48 @@ class HomeViewModel @Inject constructor(
     private var lastRefreshTime = 0L
     private var isAppInForeground = true
 
+    private val searchQueryFlow = MutableStateFlow("")
+    private var searchJob: Job? = null
+
     init {
         ProcessLifecycleOwner.get().lifecycle.addObserver(this)
-        
-        // First, listen for user changes 
+
         viewModelScope.launch {
             var previousUserId: String? = null
             preferencesStore.activeUserId.collect { userId ->
                 if (previousUserId != null && previousUserId != userId) {
-                    // User switched, force a full refresh with cleared state
                     refreshJob?.cancel()
                     resetHomeScrollPosition()
                     resetHomeFocusPosition()
-                    sections = emptyList()
-                    favorites = emptyList()
-                    discoverSections = emptyMap()
-                    error = null
-                    isLoading = true
+                    _uiState.update { it.copy(sections = emptyList(), favorites = emptyList(), discoverSections = emptyMap(), error = null, isLoading = true) }
                     fetchAndUpdateSections()
-                    isLoading = false
+                    _uiState.update { it.copy(isLoading = false) }
                     startPeriodicRefresh()
                 }
                 previousUserId = userId
             }
         }
-        
-        // Listen for preference changes
+
         viewModelScope.launch {
             preferencesStore.preferences.collect { prefs ->
-                kidsModeEnabled = prefs.kidsModeEnabled
-                homeMode = prefs.homeMode
-                dynamicTheming = prefs.dynamicTheming
-                oledMode = prefs.oledMode
                 enabledHomeSectionTypes = prefs.enabledHomeSectionTypes
                 hiddenLibrarySectionIds = prefs.hiddenLibrarySectionIds
+                _uiState.update { it.copy(
+                    kidsModeEnabled = prefs.kidsModeEnabled,
+                    homeMode = prefs.homeMode,
+                    dynamicTheming = prefs.dynamicTheming,
+                    oledMode = prefs.oledMode,
+                ) }
             }
         }
 
-        // Listen for Seerr preference changes
         viewModelScope.launch {
             seerrPreferencesStore.preferences.collect { prefs ->
-                val wasEnabled = discoverEnabled
+                val wasEnabled = _uiState.value.discoverEnabled
                 seerrPreferences = prefs
-                discoverEnabled = prefs.enabled && prefs.discoverEnabled
-                if (discoverEnabled && !wasEnabled) {
+                val nowEnabled = prefs.enabled && prefs.discoverEnabled
+                _uiState.update { it.copy(discoverEnabled = nowEnabled) }
+                if (nowEnabled && !wasEnabled) {
                     fetchDiscoverSections(prefs)
                 }
             }
@@ -168,11 +129,10 @@ class HomeViewModel @Inject constructor(
 
         viewModelScope.launch {
             offlineModeManager.offlineMode.collect { mode ->
-                val wasOffline = offlineMode != OfflineMode.ONLINE
-                offlineMode = mode
+                val wasOffline = _uiState.value.offlineMode != OfflineMode.ONLINE
+                _uiState.update { it.copy(offlineMode = mode) }
                 if (mode != OfflineMode.ONLINE && !wasOffline) {
-                    sections = emptyList()
-                    discoverSections = emptyMap()
+                    _uiState.update { it.copy(sections = emptyList(), discoverSections = emptyMap()) }
                 } else if (mode == OfflineMode.ONLINE && wasOffline) {
                     fetchAndUpdateSections()
                 }
@@ -181,46 +141,198 @@ class HomeViewModel @Inject constructor(
 
         viewModelScope.launch {
             offlineRepository.getOfflineLibrary().collect { items ->
-                offlineLibrary = items
+                _uiState.update { it.copy(offlineLibrary = items) }
             }
         }
-        
-        // Finally, load initial data
+
+        viewModelScope.launch {
+            @OptIn(FlowPreview::class)
+            searchQueryFlow
+                .debounce(400)
+                .distinctUntilChanged()
+                .collect { query ->
+                    searchJob?.cancel()
+                    if (query.isBlank()) {
+                        _uiState.update { it.copy(searchState = it.searchState.copy(jellyfinResults = emptyList(), seerrResults = emptyList(), isSearching = false)) }
+                    } else {
+                        searchJob = launch { performSearch(query) }
+                    }
+                }
+        }
+
         loadInitial()
+    }
+
+    fun onEvent(event: HomeUiEvent) {
+        when (event) {
+            is HomeUiEvent.Refresh -> refresh()
+            is HomeUiEvent.PullToRefresh -> pullToRefresh()
+            is HomeUiEvent.ToggleOfflineMode -> toggleOfflineMode()
+            is HomeUiEvent.UpdateSearchQuery -> updateSearchQuery(event.query)
+            is HomeUiEvent.ClearSearch -> clearSearch()
+            is HomeUiEvent.SelectSeerrRequestItem -> selectSeerrRequestItem(event.item)
+            is HomeUiEvent.RequestSeerrMedia -> requestSeerrMedia(event)
+            is HomeUiEvent.ClearRequestResult -> clearRequestResult()
+            is HomeUiEvent.LoadSeerrServiceDetails -> loadSeerrServiceDetails(event.mediaType)
+            is HomeUiEvent.LoadTvSeasons -> loadTvSeasons(event.tmdbId)
+            is HomeUiEvent.PrefetchSeerrDetails -> prefetchSeerrDetails(event.tmdbId, event.mediaType, event.onDone)
+        }
+    }
+
+    fun getImageUrl(itemId: String): String =
+        playbackRepository.getImageUrl(itemId, maxWidth = 400)
+
+    fun getBackdropUrl(itemId: String): String =
+        playbackRepository.getBackdropUrl(itemId, maxWidth = 1280)
+
+    fun getHomeScrollPosition(): HomeScrollPosition = homeScrollPosition
+
+    fun saveHomeScrollPosition(firstVisibleItemIndex: Int, firstVisibleItemScrollOffset: Int) {
+        homeScrollPosition = HomeScrollPosition(
+            firstVisibleItemIndex = firstVisibleItemIndex.coerceAtLeast(0),
+            firstVisibleItemScrollOffset = firstVisibleItemScrollOffset.coerceAtLeast(0),
+        )
+    }
+
+    fun resetHomeScrollPosition() {
+        homeScrollPosition = HomeScrollPosition()
+    }
+
+    fun getHomeFocusPosition(): HomeFocusPosition = homeFocusPosition
+
+    fun saveHomeFocusPosition(sectionIndex: Int, itemIndex: Int) {
+        homeFocusPosition = HomeFocusPosition(
+            sectionIndex = sectionIndex.coerceAtLeast(0),
+            itemIndex = itemIndex.coerceAtLeast(0),
+        )
+    }
+
+    fun resetHomeFocusPosition() {
+        homeFocusPosition = HomeFocusPosition()
     }
 
     private fun loadInitial() {
         viewModelScope.launch {
-            isLoading = true
-            error = null
+            _uiState.update { it.copy(isLoading = true, error = null) }
             fetchAndUpdateSections()
-            isLoading = false
+            _uiState.update { it.copy(isLoading = false) }
             startPeriodicRefresh()
         }
     }
 
-    fun refresh() {
+    private fun refresh() {
         viewModelScope.launch {
-            isLoading = true
+            _uiState.update { it.copy(isLoading = true) }
             resetHomeScrollPosition()
             resetHomeFocusPosition()
-            sections = emptyList()
-            favorites = emptyList()
-            discoverSections = emptyMap()
-            error = null
+            _uiState.update { it.copy(sections = emptyList(), favorites = emptyList(), discoverSections = emptyMap(), error = null) }
             fetchAndUpdateSections()
-            isLoading = false
+            _uiState.update { it.copy(isLoading = false) }
             startPeriodicRefresh()
         }
     }
 
-    fun onRefresh() {
+    private fun pullToRefresh() {
         viewModelScope.launch {
-            isRefreshing = true
-            error = null
+            _uiState.update { it.copy(isRefreshing = true, error = null) }
             fetchAndUpdateSections()
-            isRefreshing = false
+            _uiState.update { it.copy(isRefreshing = false) }
             startPeriodicRefresh()
+        }
+    }
+
+    private fun toggleOfflineMode() {
+        offlineModeManager.toggleManualOffline()
+    }
+
+    private fun updateSearchQuery(query: String) {
+        _uiState.update { it.copy(searchState = it.searchState.copy(query = query, isSearching = if (query.isBlank()) false else it.searchState.isSearching)) }
+        searchQueryFlow.value = query
+    }
+
+    private fun clearSearch() {
+        _uiState.update { it.copy(searchState = HomeSearchState()) }
+        searchQueryFlow.value = ""
+    }
+
+    private fun selectSeerrRequestItem(item: SeerrSearchItem?) {
+        _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(requestItem = item)) }
+    }
+
+    private fun requestSeerrMedia(event: HomeUiEvent.RequestSeerrMedia) {
+        viewModelScope.launch {
+            _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(result = DiscoverRequestResult(isLoading = true))) }
+            seerrRepository.requestMedia(
+                mediaType = event.item.mediaType,
+                tmdbId = event.item.id,
+                seasons = event.seasons,
+                serverId = event.serverId,
+                profileId = event.profileId,
+                rootFolder = event.rootFolder,
+                tags = event.tags,
+            ).onSuccess {
+                _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(result = DiscoverRequestResult(success = true))) }
+            }.onFailure { throwable ->
+                _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(result = DiscoverRequestResult(error = throwable.message ?: "Request failed"))) }
+            }
+        }
+    }
+
+    private fun clearRequestResult() {
+        _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(result = null)) }
+    }
+
+    private fun loadSeerrServiceDetails(mediaType: String) {
+        viewModelScope.launch {
+            _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(isLoadingServices = true)) }
+            try {
+                if (mediaType == "movie") {
+                    seerrRepository.getServiceRadarrServers().onSuccess { servers ->
+                        val details = servers.mapNotNull { server ->
+                            seerrRepository.getServiceRadarrDetail(server.id).getOrNull()
+                        }
+                        _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(radarrServers = details)) }
+                    }
+                } else {
+                    seerrRepository.getServiceSonarrServers().onSuccess { servers ->
+                        val details = servers.mapNotNull { server ->
+                            seerrRepository.getServiceSonarrDetail(server.id).getOrNull()
+                        }
+                        _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(sonarrServers = details)) }
+                    }
+                }
+            } finally {
+                _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(isLoadingServices = false)) }
+            }
+        }
+    }
+
+    private fun loadTvSeasons(tmdbId: Int) {
+        viewModelScope.launch {
+            _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(tvSeasons = emptyList())) }
+            seerrRepository.getTvDetails(tmdbId).onSuccess { details ->
+                _uiState.update { it.copy(seerrRequestState = it.seerrRequestState.copy(tvSeasons = details.seasons.filter { it.seasonNumber > 0 })) }
+            }
+        }
+    }
+
+    private fun prefetchSeerrDetails(tmdbId: Int, mediaType: String, onDone: () -> Unit) {
+        viewModelScope.launch {
+            try {
+                coroutineScope {
+                    if (mediaType == "movie") {
+                        seerrRepository.getMovieDetails(tmdbId)
+                    } else {
+                        seerrRepository.getTvDetails(tmdbId)
+                    }
+                    val type = if (mediaType == "movie") com.raulshma.jellyplay.core.model.MediaType.MOVIE else com.raulshma.jellyplay.core.model.MediaType.SERIES
+                    launch { seerrRepository.getRatings(tmdbId, mediaType) }
+                    launch { seerrRepository.getRecommendations(tmdbId, type) }
+                    launch { seerrRepository.getSimilar(tmdbId, type) }
+                }
+            } catch (_: Exception) {
+            }
+            onDone()
         }
     }
 
@@ -248,9 +360,7 @@ class HomeViewModel @Inject constructor(
                         }
                     } else fetchedSections
 
-                    if (this@HomeViewModel.sections != filteredSections) {
-                        this@HomeViewModel.sections = filteredSections
-                    }
+                    _uiState.update { it.copy(sections = filteredSections) }
 
                     val continueWatching = filteredSections
                         .find { it.type == HomeSectionType.CONTINUE_WATCHING }
@@ -270,22 +380,19 @@ class HomeViewModel @Inject constructor(
                                 val filteredFavorites = result.items.filter { item ->
                                     isAllowedForKids(item, prefs.kidsModeMaxRating)
                                 }
-                                if (this@HomeViewModel.favorites != filteredFavorites) {
-                                    this@HomeViewModel.favorites = filteredFavorites
-                                }
+                                _uiState.update { it.copy(favorites = filteredFavorites) }
                             }
                     }
 
-                    error = null
+                    _uiState.update { it.copy(error = null) }
                 }
-                .onFailure {
-                    if (sections.isEmpty()) {
-                        error = it.message ?: "${it::class.simpleName}"
+                .onFailure { throwable ->
+                    if (_uiState.value.sections.isEmpty()) {
+                        _uiState.update { s -> s.copy(error = throwable.message ?: "${throwable::class.simpleName}") }
                     }
                 }
 
-            // Fetch discover sections if enabled
-            if (discoverEnabled) {
+            if (_uiState.value.discoverEnabled) {
                 fetchDiscoverSections(seerrPreferences)
             }
         } finally {
@@ -326,9 +433,7 @@ class HomeViewModel @Inject constructor(
             }
         }
 
-        if (newSections != discoverSections) {
-            discoverSections = newSections
-        }
+        _uiState.update { it.copy(discoverSections = newSections) }
     }
 
     private fun startPeriodicRefresh() {
@@ -363,184 +468,10 @@ class HomeViewModel @Inject constructor(
         refreshJob?.cancel()
     }
 
-    private fun isAllowedForKids(item: MediaItem, maxRating: String): Boolean {
-        if (item.officialRating == null) return true
-        val kidRatings = listOf("G", "TV-Y", "TV-Y7", "TV-G", "PG", "TV-PG")
-        val maxIndex = kidRatings.indexOf(maxRating)
-        val itemIndex = kidRatings.indexOf(item.officialRating)
-        return if (itemIndex >= 0 && maxIndex >= 0) itemIndex <= maxIndex else true
-    }
-
-    fun getImageUrl(itemId: String): String =
-        playbackRepository.getImageUrl(itemId, maxWidth = 400)
-
-    fun getBackdropUrl(itemId: String): String =
-        playbackRepository.getBackdropUrl(itemId, maxWidth = 1280)
-
-    fun toggleOfflineMode() {
-        offlineModeManager.toggleManualOffline()
-    }
-
-    fun getHomeScrollPosition(): HomeScrollPosition = homeScrollPosition
-
-    fun saveHomeScrollPosition(firstVisibleItemIndex: Int, firstVisibleItemScrollOffset: Int) {
-        homeScrollPosition = HomeScrollPosition(
-            firstVisibleItemIndex = firstVisibleItemIndex.coerceAtLeast(0),
-            firstVisibleItemScrollOffset = firstVisibleItemScrollOffset.coerceAtLeast(0),
-        )
-    }
-
-    fun resetHomeScrollPosition() {
-        homeScrollPosition = HomeScrollPosition()
-    }
-
-    fun getHomeFocusPosition(): HomeFocusPosition = homeFocusPosition
-
-    fun saveHomeFocusPosition(sectionIndex: Int, itemIndex: Int) {
-        homeFocusPosition = HomeFocusPosition(
-            sectionIndex = sectionIndex.coerceAtLeast(0),
-            itemIndex = itemIndex.coerceAtLeast(0),
-        )
-    }
-
-    fun resetHomeFocusPosition() {
-        homeFocusPosition = HomeFocusPosition()
-    }
-
-    // ── Seerr request support ──
-
-    private val _requestResult = MutableStateFlow<DiscoverRequestResult?>(null)
-    val requestResult: StateFlow<DiscoverRequestResult?> = _requestResult.asStateFlow()
-
-    private val _radarrServers = MutableStateFlow<List<SeerrRadarrServiceDetail>>(emptyList())
-    val radarrServers: StateFlow<List<SeerrRadarrServiceDetail>> = _radarrServers.asStateFlow()
-
-    private val _sonarrServers = MutableStateFlow<List<SeerrSonarrServiceDetail>>(emptyList())
-    val sonarrServers: StateFlow<List<SeerrSonarrServiceDetail>> = _sonarrServers.asStateFlow()
-
-    private val _isLoadingSeerrServices = MutableStateFlow(false)
-    val isLoadingSeerrServices: StateFlow<Boolean> = _isLoadingSeerrServices.asStateFlow()
-
-    private val _tvSeasons = MutableStateFlow<List<SeerrSeason>>(emptyList())
-    val tvSeasons: StateFlow<List<SeerrSeason>> = _tvSeasons.asStateFlow()
-
-    fun requestSeerrMedia(
-        item: SeerrSearchItem,
-        seasons: List<Int>? = null,
-        serverId: Int? = null,
-        profileId: Int? = null,
-        rootFolder: String? = null,
-        tags: List<Int>? = null,
-    ) {
-        viewModelScope.launch {
-            _requestResult.value = DiscoverRequestResult(isLoading = true)
-            seerrRepository.requestMedia(
-                mediaType = item.mediaType,
-                tmdbId = item.id,
-                seasons = seasons,
-                serverId = serverId,
-                profileId = profileId,
-                rootFolder = rootFolder,
-                tags = tags,
-            ).onSuccess {
-                _requestResult.value = DiscoverRequestResult(success = true)
-            }.onFailure {
-                _requestResult.value = DiscoverRequestResult(error = it.message ?: "Request failed")
-            }
-        }
-    }
-
-    fun clearRequestResult() {
-        _requestResult.value = null
-    }
-
-    fun loadSeerrServiceDetails(mediaType: String) {
-        viewModelScope.launch {
-            _isLoadingSeerrServices.value = true
-            try {
-                if (mediaType == "movie") {
-                    seerrRepository.getServiceRadarrServers().onSuccess { servers ->
-                        val details = servers.mapNotNull { server ->
-                            seerrRepository.getServiceRadarrDetail(server.id).getOrNull()
-                        }
-                        _radarrServers.value = details
-                    }
-                } else {
-                    seerrRepository.getServiceSonarrServers().onSuccess { servers ->
-                        val details = servers.mapNotNull { server ->
-                            seerrRepository.getServiceSonarrDetail(server.id).getOrNull()
-                        }
-                        _sonarrServers.value = details
-                    }
-                }
-            } finally {
-                _isLoadingSeerrServices.value = false
-            }
-        }
-    }
-
-    fun loadTvSeasons(tmdbId: Int) {
-        viewModelScope.launch {
-            _tvSeasons.value = emptyList()
-            seerrRepository.getTvDetails(tmdbId).onSuccess { details ->
-                _tvSeasons.value = details.seasons.filter { it.seasonNumber > 0 }
-            }
-        }
-    }
-
-    // ── Home search ──
-
-    var searchQuery by mutableStateOf("")
-        private set
-    var searchJellyfinResults by mutableStateOf<List<MediaItem>>(emptyList())
-        private set
-    var searchSeerrResults by mutableStateOf<List<SeerrSearchItem>>(emptyList())
-        private set
-    var isSearching by mutableStateOf(false)
-        private set
-
-    private val searchQueryFlow = MutableStateFlow("")
-    private var searchJob: Job? = null
-
-    init {
-        viewModelScope.launch {
-            @OptIn(FlowPreview::class)
-            searchQueryFlow
-                .debounce(400)
-                .distinctUntilChanged()
-                .collect { query ->
-                    searchJob?.cancel()
-                    if (query.isBlank()) {
-                        searchJellyfinResults = emptyList()
-                        searchSeerrResults = emptyList()
-                        isSearching = false
-                    } else {
-                        searchJob = launch { performSearch(query) }
-                    }
-                }
-        }
-    }
-
-    fun updateSearchQuery(query: String) {
-        searchQuery = query
-        searchQueryFlow.value = query
-        if (query.isBlank()) {
-            isSearching = false
-        }
-    }
-
-    fun clearSearch() {
-        searchQuery = ""
-        searchQueryFlow.value = ""
-        searchJellyfinResults = emptyList()
-        searchSeerrResults = emptyList()
-        isSearching = false
-    }
-
     private suspend fun performSearch(query: String) {
-        isSearching = true
+        _uiState.update { it.copy(searchState = it.searchState.copy(isSearching = true)) }
         try {
-            kotlinx.coroutines.coroutineScope {
+            coroutineScope {
                 val jellyfinDeferred = async { mediaRepository.search(query, limit = 8) }
                 val seerrDeferred = async {
                     try {
@@ -556,60 +487,28 @@ class HomeViewModel @Inject constructor(
                     }
                 }
                 jellyfinDeferred.await().onSuccess { result ->
-                    searchJellyfinResults = result.items
+                    _uiState.update { it.copy(searchState = it.searchState.copy(jellyfinResults = result.items)) }
                 }
                 seerrDeferred.await()?.let { results ->
-                    searchSeerrResults = results
+                    _uiState.update { it.copy(searchState = it.searchState.copy(seerrResults = results)) }
                 } ?: run {
-                    searchSeerrResults = emptyList()
+                    _uiState.update { it.copy(searchState = it.searchState.copy(seerrResults = emptyList())) }
                 }
             }
         } catch (e: CancellationException) {
             throw e
         } catch (_: Exception) {
-            searchJellyfinResults = emptyList()
-            searchSeerrResults = emptyList()
+            _uiState.update { it.copy(searchState = it.searchState.copy(jellyfinResults = emptyList(), seerrResults = emptyList())) }
         } finally {
-            isSearching = false
+            _uiState.update { it.copy(searchState = it.searchState.copy(isSearching = false)) }
         }
     }
 
-    fun prefetchSeerrDetails(tmdbId: Int, mediaType: String, onDone: () -> Unit) {
-        viewModelScope.launch {
-            try {
-                coroutineScope {
-                    if (mediaType == "movie") {
-                        seerrRepository.getMovieDetails(tmdbId)
-                    } else {
-                        seerrRepository.getTvDetails(tmdbId)
-                    }
-                    val type = if (mediaType == "movie") com.raulshma.jellyplay.core.model.MediaType.MOVIE else com.raulshma.jellyplay.core.model.MediaType.SERIES
-                    launch { seerrRepository.getRatings(tmdbId, mediaType) }
-                    launch { seerrRepository.getRecommendations(tmdbId, type) }
-                    launch { seerrRepository.getSimilar(tmdbId, type) }
-                }
-            } catch (_: Exception) {
-            }
-            onDone()
-        }
+    private fun isAllowedForKids(item: com.raulshma.jellyplay.core.model.MediaItem, maxRating: String): Boolean {
+        if (item.officialRating == null) return true
+        val kidRatings = listOf("G", "TV-Y", "TV-Y7", "TV-G", "PG", "TV-PG")
+        val maxIndex = kidRatings.indexOf(maxRating)
+        val itemIndex = kidRatings.indexOf(item.officialRating)
+        return if (itemIndex >= 0 && maxIndex >= 0) itemIndex <= maxIndex else true
     }
 }
-
-@Immutable
-data class HomeScrollPosition(
-    val firstVisibleItemIndex: Int = 0,
-    val firstVisibleItemScrollOffset: Int = 0,
-)
-
-@Immutable
-data class HomeFocusPosition(
-    val sectionIndex: Int = 0,
-    val itemIndex: Int = 0,
-)
-
-@Immutable
-data class DiscoverRequestResult(
-    val isLoading: Boolean = false,
-    val success: Boolean? = null,
-    val error: String? = null,
-)
